fix(onboard): height, weight 자료형 수정

+++ sweet-choice/src/main/java/com/codeon/sweet_choice/dto/SignupRequestDto.java
@@ -19,8 +19,8 @@ public class SignupRequestDto {
     // 사용자 상태
     private String userType;
 
-    private float height;
-    private float weight;
+    private Integer height;
+    private Integer weight;
 
     private String sex;
 }
